+++ Engine/src/engine/ui/util/RectTransformUtil.cpp
@@ -11,38 +11,38 @@ namespace se::ui::util
         Rect ret = {};
         if (transform.anchors.left > 0.5f)
         {
-            ret.topLeft.x = (parentTransform.rect.topLeft.x + transform.anchors.left * parentWidth) - transform.minX;
+            ret.topLeft.x = static_cast<int>((parentTransform.rect.topLeft.x + transform.anchors.left * parentWidth) - transform.minX);
         }
         else
         {
-            ret.topLeft.x = (parentTransform.rect.topLeft.x + transform.anchors.left * parentWidth) + transform.minX;
+            ret.topLeft.x = static_cast<int>((parentTransform.rect.topLeft.x + transform.anchors.left * parentWidth) + transform.minX);
         }
 
         if (transform.anchors.top > 0.5f)
         {
-            ret.topLeft.y = (parentTransform.rect.topLeft.y + transform.anchors.top * parentHeight) - transform.minY;
+            ret.topLeft.y = static_cast<int>((parentTransform.rect.topLeft.y + transform.anchors.top * parentHeight) - transform.minY);
         }
         else
         {
-            ret.topLeft.y = (parentTransform.rect.topLeft.y + transform.anchors.top * parentHeight) + transform.minY;
+            ret.topLeft.y = static_cast<int>((parentTransform.rect.topLeft.y + transform.anchors.top * parentHeight) + transform.minY);
         }
 
         if (transform.anchors.right > 0.5f)
         {
-            ret.size.x = (parentTransform.rect.topLeft.x + transform.anchors.right * parentWidth) - transform.maxX - ret.topLeft.x;
+            ret.size.x = static_cast<int>((parentTransform.rect.topLeft.x + transform.anchors.right * parentWidth) - transform.maxX - ret.topLeft.x);
         }
         else
         {
-            ret.size.x = (parentTransform.rect.topLeft.x + transform.anchors.right * parentWidth) + transform.maxX - ret.topLeft.x;
+            ret.size.x = static_cast<int>((parentTransform.rect.topLeft.x + transform.anchors.right * parentWidth) + transform.maxX - ret.topLeft.x);
         }
 
         if (transform.anchors.bottom > 0.5f)
         {
-            ret.size.y = (parentTransform.rect.topLeft.y + transform.anchors.bottom * parentHeight) - transform.maxY - ret.topLeft.y;
+            ret.size.y = static_cast<int>((parentTransform.rect.topLeft.y + transform.anchors.bottom * parentHeight) - transform.maxY - ret.topLeft.y);
         }
         else
         {
-            ret.size.y = (parentTransform.rect.topLeft.y + transform.anchors.bottom * parentHeight) + transform.maxY - ret.topLeft.y;
+            ret.size.y = static_cast<int>((parentTransform.rect.topLeft.y + transform.anchors.bottom * parentHeight) + transform.maxY - ret.topLeft.y);
         }
         return ret;
     }
